pyproject.toml
[project]
name = "photo-album-rag"
version = "0.1.0"
description = "A fully local RAG system for your photo library using contextual retrieval"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Han", email = "bearbearyu@gmail.com" }]
keywords = ["rag", "photos", "llm", "contextual-retrieval", "local-ai"]

dependencies = [
    # Core RAG stack
    "chromadb>=0.5.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=3.0.0",
    # Photo processing
    "pillow>=10.0.0",
    "pillow-heif>=0.16.0",   # HEIC/HEIF support for iPhone photos
    "exifread>=3.0.0",
    # Geocoding
    "geopy>=2.4.0",
    # Local LLM / vision model via Ollama
    "ollama>=0.3.0",
    # CLI / UX
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
# Better reranking model (larger, more accurate)
rerank-large = [
    "sentence-transformers>=3.0.0",   # already a core dep, but signals intent
]
# Image-similarity search via CLIP embeddings
clip = [
    "open-clip-torch>=2.26.0",
    "torch>=2.0.0",
]
# Simple web UI for browsing results with thumbnails
ui = [
    "streamlit>=1.35.0",
]
# Everything
all = [
    "photo-album-rag[clip,ui]",
]
# Development tools
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
# After `uv sync`, these commands become available in the venv
photo-ingest  = "photo_rag.ingest:main"
photo-search  = "photo_rag.retrieve:main"
photo-query   = "photo_rag.query:main"
photo-app     = "photo_rag.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/photo_rag"]

# ── uv ────────────────────────────────────────────────────────────────────────

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

# ── Ruff (linter + formatter) ─────────────────────────────────────────────────

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]   # line length handled by formatter

[tool.ruff.format]
quote-style = "double"

# ── Mypy ──────────────────────────────────────────────────────────────────────

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true   # many ML libs lack stubs

# ── Pytest ────────────────────────────────────────────────────────────────────

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
